fix sales order

+++ database/migrations/2022_11_22_082344_create_sales_order_details_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->integer('keterangan')->nullable();
 
 
-            $table->integer('salesorder_id')->unsigned();
+            $table->bigInteger('salesorder_id')->unsigned();
             $table->foreign('salesorder_id')->references('id')->on('sales_orders')
                   ->onDelete('cascade');
